backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.4
celery==5.4.0
redis==5.2.1
django-ratelimit==4.1.0
django-prometheus==2.3.1
structlog==24.4.0
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-django==0.49b2
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==5.0.0
